Refactor: split aml emitters into separate files
