+++ shared/src/main/scala/amf/plugins/document/vocabularies/emitters/instances/ReferenceEmitter.scala
@@ -1,0 +1,21 @@
+package amf.plugins.document.vocabularies.emitters.instances
+import amf.core.emitter.BaseEmitters.MapEntryEmitter
+import amf.core.emitter.{EntryEmitter, SpecOrdering}
+import amf.core.model.document.BaseUnit
+import amf.core.parser.Position
+import amf.core.parser.Position.ZERO
+import org.yaml.model.YDocument.EntryBuilder
+
+case class ReferenceEmitter(reference: BaseUnit, ordering: SpecOrdering, aliases: Map[String, (String, String)])
+    extends EntryEmitter {
+
+  override def emit(b: EntryBuilder): Unit = {
+    aliases.get(reference.id) match {
+      case Some((alias, location)) =>
+        MapEntryEmitter(alias, location).emit(b)
+      case _ => // TODO: emit violation
+    }
+  }
+
+  override def position(): Position = ZERO
+}
